V1-D-811612 lots of refactoring,

+++ persistence/src/main/java/org/openstack/atlas/service/domain/services/impl/SslTerminationServiceImpl.java
@@ -69,10 +69,12 @@ public class SslTerminationServiceImpl extends BaseService implements SslTermina
 
     @Transactional
     @Override
-    public boolean deleteSslTermination(Integer lid, Integer accountId) throws EntityNotFoundException, ImmutableEntityException, UnprocessableEntityException {
+    public boolean deleteSslTermination(Integer lid, Integer accountId) throws EntityNotFoundException, ImmutableEntityException, UnprocessableEntityException, BadRequestException {
         LoadBalancer dbLoadBalancer = loadBalancerRepository.getByIdAndAccountId(lid, accountId);
         isLbActive(dbLoadBalancer);
-
+        if (!dbLoadBalancer.hasSsl()) {
+            throw new BadRequestException("SSL Termination to be deleted not found on loadbalancer: " + dbLoadBalancer.getId());
+        }
         return sslTerminationRepository.removeSslTermination(lid, accountId);
     }
